+++ src/app/layout.js
@@ -1,6 +1,5 @@
 import { Inter } from "next/font/google";
 import "./globals.css";
-import Navbar from "@/components/Navbar";
 import { Box } from "@mui/material";
 
 const inter = Inter({ subsets: ["latin"] });
